Preliminary/Pseudo Auto Code
This is some very basic auto code that is pretty much just comments. It does however give a nice list of things we need for some successful auto and lays out a bit, as well as providing a place for us to write the auto itself when the time comes.